services/crowdflow-agent/pyproject.toml
[project]
name = "crowdflow-agent"
version = "0.1.0"
description = "A crowd flow analysis agent that uses computer vision, YOLOv3 detection, and custom ML models to predict crowd density patterns and generate flow predictions."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0,<2.0.0",
    "google-cloud-aiplatform[adk,agent_engine]>=1.100.0",
    "google-cloud-vision>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "python-dotenv>=1.0.0",
    "pathlib",
    "asyncio",
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "tensorflow>=2.10.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false 